Simple project using python in UAE  Space Agency
In this simple project, I used the Python programming language to create a simple system that simulates some of the theoretical and practical tests to examine applicants for the Emirates Space Agency. Some of these tests include a system examination using the Tikenter library, a body mass examination, and many other tests.